perf(eval-tasks): single-statement reconciler requeue
_requeue_and_drop fired one UPDATE per eval-config group (N configs -> N
statements). Collapse to a single UPDATE across all requeued ids using a
Case/When on custom_eval_config_id so each entry still gets its own config's
config_hash. Chunk the combined id list at _REQUEUE_CHUNK (10k) to stay well
under Postgres's 65535-parameter limit. Behavior is unchanged: same rows
requeued, each with the correct hash, error/skipped_reason cleared.

Stress test S10: asserts the requeue statement count and requeue-count parity.
